Integrate memoization syntax
Adds AST nodes for memoization and handle this in the registry. A direct
port from the current Ruby and JavaScript setup. There is probably a
more elegant way to do this by passing the context down the tree during
evaluation rather than bundling everything into the registry, but for
now, let’s just follow the current working codebase as-is before
introducing any large changes.