Tidy up registry tests and add unique cycle test
Replace checks on expression tree structure with basic string checks in
the Registry evaluation tests. All the registry test needs to do is
assert the returned expression can be flattened to the expected string.
More intricate tests of the tree structure should be done with tests for
the syntax nodes. Some additional details added to the expansion test as
a starting point.