add release Github action
This action is trigger by a tag being pushed to the remote, in the form of `v1.23.456`, which is used for identifying the version published on NuGet. Assembly*Version properties in the .csproj still have their usual meanings in the MSBuild workflow. 

Once built and tests pass, it is packaged up and published to NuGet using the NUGET_API_KEY secret from the repository