refactor Expansion class to avoid the Term/Tail duality
This removes the possibility of creating an Expression object with an invalid state that would crash CollectAtoms()

This refactor shook up a lot of the tests, as they were inspecting the internals directly. These tests have been replaced by object equality tests where appropriate which seem a little cleaner.

Non-atom Expression types include a convenience initializer which allows the expression and the atom it contains be initialized in one call.